beaglebone-yocto: move kernel and dtb to packagegroup-core-boot
Appending IMAGE_INSTALL with the kernel and device tree is a bad idea
because it will bring these packages to every images, like a container
image or sysroot image to pivot to, where the kernel artifacts are
not wanted.

Instead use the MACHINE_ESSENTIAL_EXTRA_RDEPENDS which results
in having the kernel artifacts where they should be, part of
packagegroup-core-boot.

Signed-off-by: Vivien Didelot <vdidelot@pbsc.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>