poky-tiny: set QB_DEFAULT_FSTYPE correctly
QB_DEFAULT_FSTYPE is used literally as the image extension in the testimage
code, so this shouldn't be set as IMAGE_FSTYPES because that variable is a
list of image types, so if IMAGE_FSTYPES is extended and testimage ran it
will try to look for an image called e.g. "core-image-minimal-cpio.gz ext4".

Instead, set QB_DEFAULT_FSTYPE explicitly to cpio.gz.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>